@@ -4,13 +4,16 @@
 	"description": "Check if a JavaScript environment has a DOM",
 	"license": "MIT",
 	"repository": "sindresorhus/has-dom",
+	"funding": "https://github.com/sponsors/sindresorhus",
 	"author": {
 		"name": "Sindre Sorhus",
 		"email": "sindresorhus@gmail.com",
-		"url": "sindresorhus.com"
+		"url": "https://sindresorhus.com"
 	},
+	"type": "module",
+	"exports": "./index.js",
 	"engines": {
-		"node": ">=6"
+		"node": ">=12"
 	},
 	"scripts": {
 		"test": "xo && ava"
@@ -30,8 +33,8 @@
 		"check"
 	],
 	"devDependencies": {
-		"ava": "*",
-		"xo": "*"
+		"ava": "^3.15.0",
+		"xo": "^0.39.1"
 	},
 	"xo": {
 		"envs": [
