@@ -1,6 +1,5 @@
-'use strict';
-module.exports = () => {
+export default function hasDom() {
 	return typeof window !== 'undefined' &&
 		typeof document !== 'undefined' &&
 		typeof document.createElement === 'function';
-};
+}
